requirements.txt
etcd3==0.12.0
Flask==2.1.2
openstacksdk==0.61.0
PyYAML==6.0
protobuf==3.20.1
apscheduler
